Dispatch to drawing by id
This update uses drawings/by-id to return the drawing definition used to
render a project component. The purpose of this is largely for routing
--- a url will look somehting like "/drawings/86" which we'll break into
a map like {:page :drawing, :query {:id 86}}. :page will let us dispatch
to the correct component and the query param will be fed in as props.