Create canvas demo
It's evident that SVG isn't going to cut it when it comes to rendering
10,000 lines. In these sorts of times, we reach for <canvas>:

  https://bocoup.com/blog/smoothly-animate-thousands-of-points-with-html5-canvas-and-d3

Canvas has a very stateful and context-heavy api as it is. Wrapping that
in React and calling lifecycle methods on the context object requires
some finesse, and doing that in a clojure wrapper requires it even more
so.

For that reason, I decided to give Rum the boot and switch to the
dead-simple, no macro, no magic brutha:

  https://github.com/weavejester/brutha

Which has fairly one-to-one mapping for those callbacks, allowing us to
hook into on componentDidMount and componentWillRecieveProps and all
that. All of our components were defined using the sablono-style
functions. To use them with brutha, we just needed to call the
sablono/html function to build components out of them.

This also seemed like a decent opportunity to put all of the
component-related code hanging out in core into a components namespace.

There's a lot in there around points and line generation that will
likely be put in the drawings namespace, or even extracted into
general-purpose shape stuff.

Either way, it gets us to a point where we can do some basic shape
drawing on mount of a canvas component.