Autosize the drawing to fit the page
Fun fact: I used to be on a team with react-virtualized's Brian Vaughn.
Guy hated React with a passion.  He mostly made "your mom" jokes to our
CTO. Three years later, the guy is on the React core team and I'm using
a(n excellent) library for observing an elment's size.

I think there still needs to be a lot of work around how the element in
integrated, but this commit adds the <AutoSizer> component from
react-virtualized to provide the canvas component with the size of its
container. This is nice! We can rely on CSS to do the work of measuring
the size available to us to draw on after laying out the bottom panel
with the description of the work. I'm imagining the ideal api for
something like this is is to call (with-size <component>) and doing some
cloneElement magic to inject the dimensions into the component as props,
but that's going to take a metric ton of fuss to get right and it's
friday. Cheers.