Build out helper functions for canvas drawing
Didn't realize it for a while, but I ended up buliding a partial
implementation of monet:

  https://github.com/rm-hull/monet

The main benefit of wrapping the javascript canvas apis is to:

- normalize method calls and setting properties to the same general
  calling syntax

- by providing the drawing context as the first argument to all of these
  functions, we can use the thread-first (->) macro to pass the context
  in as the first argument for a chain of updates

It's that second benefit that I'm really after.

Of course, all of this might go out the window because the drawing APIs
have a huge drawback: they can't draw crisp, pixelated, non-anti-aliased
lines.

Maybe in a next version of all of this, I'll end up with a different
render target. It _is_ possible to make those crisp crisp lines, but
only if you draw along a pixel's edge (for straight lines):

  http://stackoverflow.com/questions/7530593/html5-canvas-and-line-width/7531540#7531540

or implement your own Bresenham for drawing diagonals:

  http://stackoverflow.com/questions/25277023/complete-solution-for-drawing-1-pixel-line-on-html5-canvas/25279040#25279040

Honestly, that sounds like a lot of fun, so I'm probably going to do it
anyway. But it might be interesting to implement it as a multimethod,
where `draw-line` can be implemented based on some :crisp flag.