Bulid canvas image from shape definitions
One nice thing I had from back in the SVG days was a way of keeping the
definition of lines separate from the svg representing those lines. When
it came to sablono DOM trees and how close my "description map of
a line" data structures looked, at a certain point I thougth "whatever"
and just started generating SVG in lewitt.drawings and called it a day.

It was better before though! What the drawing definitions should do is
return a list of shapes given a canvas size, and our canvas abstraction
should take that list of shapes and output a drawing.

That's the thinking behind the draw-from-shapes function, which will
dispatch on the :kind field of the shape, and invoke the correct canvas
methods to render it to a canvas.

This is nice because the drawing description can remain very high-level
(just output a list of shapes) and the canvas drawing layer can remain
very flexible (swap out the `draw-line` function with a Bresenham
implementation!). I'm thinking the target can change depending on
browser support or speed.