Add bresenham line renderer
This commit adds a (slow!) bresenham line renderer. It takes something
on the order of 7 seconds to render 10000 lines for drawing 86.

The *look* is amazing --- the jagged lines are by far one of the nicest
aesthetics you could go for. I'll have to profile it to see where the
code is spending most of its time. There are a few posibilities for what
could speed it up:

- Drawing the lines could happen asynchronously, either in a web worker
or through some chunking of the work so that the thread doesn't get tied
up on buliding this drawing. This would be really nice in general for
performing the work of the drawing, and would ensure the app remains
fast no matter how inefficient the drawing algorithm is, which if this
thing ever intends to have user-submitted input for drawing
descriptions, it may very well have to be.

- Draw the whole thing incrementally. Do as much work in an animation
frame as possible, and when the budget is up, render that to the screen.
It might take 7 seconds to render, but that might look really
interesting if you get to watch it bulid up rather than beachball during
that time.

- Drop down to implementing the canvas renderer in pure js. It might be
much faster to convert the vector of shape maps to an array of shape
objects, and handle all line drawing in JS, rather than continuously
jumping between the cljs and js contexts. This is a major assumption
though, and will need both testing and a review of the generated code.