Add drawing 142 (not straight lines)
Drawing 142 has a grid with "not straight" lines parallel to the grid
lines. As a part of the layer that generates the image, a new shape kind
is introduced, the :not-straight-line type. Right now, I'm going to let
the renderer determine what "not straight means", how it'll handle
the total variance. There are algorithms for this:

  - http://webhome.cs.uvic.ca/~blob/publications/zainabcag.pdf
  - https://github.com/balint42/comic.js
  - https://29a.ch/2010/2/10/hand-drawn-lines-algorithm-javascript-canvas-html5

Naturally they are very imperative. I'm beginning to think that this
project needs to have the canvas renderer implemented in javascript, and
cljs can call into it.