Set-pixel function allows specifying the color
I'd previously hard-coded the set-pixel function to blue just to get it
running. This makes it so that you pass in two values to set-pixel:

- a vector of the rgb values
- a vector representing the position

I figure it makes sense to make color the first argument to this
function since you're more likely to curry the color than the position.
As a vector, we can destructure in the arguments to the r g b vars.