Add "not-straight-lines" to drawing primitives
"not straight lines" seems to be a pretty common tool in LeWitt's work.
And a very interesting thing to implement functionally. So far, my
thinking is that a renderer should have some concept of a "not straight
line", which can take into account the line's thickness, the canvas
size, and the level of variation that makes sense for the medium.
The canvas renderer here draws a Bresenham line, as the jaggedness is
makes for a nice computer-y feel. The machine renderer may opt to draw
more quickly and erratically, taking into account the physical setup,
and that a quickly drawn, jittery line is more likely to not be
straight.